% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors}
\alias{colors}
\title{Simulated data set generated under the minimal repair model}
\format{A list with 3 variables:
\describe{
  \item{main}{First colors}
  \item{secondary}{Secondary colors}
  \item{grey}{gray colors}
}}
\usage{
colors
}
\description{
This contains colors used in pallete of my Phd thesis.
}
\keyword{datasets}
