% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mythesis_pal}
\alias{mythesis_pal}
\title{Phd thesis color palettes}
\usage{
mythesis_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{the palette to be used, options are \code{c('main', 'secondary', 'grey')}}

\item{reverse}{boolean, represents the order of the palette}

\item{...}{Named argument to modify the palette}
}
\description{
Phd thesis color palettes
}
