% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_mythesis}
\alias{scale_color_mythesis}
\title{Phd thesis scale colors for ggplot2}
\usage{
scale_color_mythesis(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{the palette to be used, options are \code{c('main', 'secondary', 'grey')}}

\item{discrete}{boolean, the scale is discrete? \code{TRUE} if it is.}

\item{reverse}{boolean, represents the order of the palette}

\item{...}{Named argument to modify the \code{scale_color_} fucntion}
}
\description{
Phd thesis scale colors for ggplot2
}
