% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_mythesis}
\alias{theme_mythesis}
\title{Phd thesis theme for ggplot2}
\usage{
theme_mythesis(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Phd thesis theme for ggplot2
}
